Moved Cassie UDP-LCM translation to it's own file.
