+++ examples/Cassie/networking/BUILD.bazel
@@ -11,6 +11,16 @@ load(
 )
 
 cc_library(
+  name = "udp_lcm_translator",
+  srcs = ["udp_lcm_translator.cc",],
+  hdrs = ["udp_lcm_translator.h"],
+  deps = [
+    "//examples/Cassie/datatypes:cassie_inout_types",
+    "//lcmtypes:lcmt_robot",
+  ]
+)
+
+cc_library(
   name = "cassie_udp_pub_sub",
   srcs = ["udp_serializer.cc",
           "cassie_udp_subscriber.cc",
@@ -36,7 +46,8 @@ cc_library(
     "//lcmtypes:lcmt_robot",
     "//multibody:utils",
     "//attic/multibody:utils",
-    ":simple_cassie_udp_subscriber"
+    ":simple_cassie_udp_subscriber",
+    ":udp_lcm_translator",
   ]
 )
